2023.1
 # New release naming convention

Please note from this release onwards we will be using the YYYY.M.rev
format for release names/versions. This change is intended to allow for greater
clarity regarding deprecations or upcoming breaking changes.

 # Updates

* asio fork updated to include latest upstream changes
* Fix potential use-after-free in WinHTTP callback
* InvalidUser is now part of the UserNotAuthenticatedError condition
* Improvements to the safety of modfile decompression
* Cache logic improvements for mod subscriptions and unsubscriptions
* SubmitModChangesAsync now has support for editing the mod logo
* Documentation updates
* Modio::AuthenticationParams now has a flag to request the SDK URLEncodes the auth token

 # Deprecations

* The public progress members on ModProgressInfo have been deprecated in favour of the
following three methods:
  * `Modio::FileSize GetCurrentProgress(EModProgressState State) const`
  * `Modio::FileSize GetTotalProgress(EModProgressState State) const`
  * `EModProgressState GetCurrentState() const`

  This change means that users no longer have to intuit the state of a mod upload
or download by inspecting the values of the ModProgressInfo object and doing a
comparison between current and total values. It also means that the object is now
explicit with an operation's state rather than relying on context from
ModManagementEvents or similar.

  As a result of these changes, in a forthcoming release, a file
transfer that is starting, but relies on a REST API call to begin
reporting progress (for example, a file transfer that requires us to
fetch the file size from the server) will result in a ModProgressInfo
object being returned with a state of Initializing while that initial
request is performed. This will be a
breaking change from the current behaviour where the SDK simply doesn't
return a ModProgressInfo object until that file size has been retrieved,
and will be documented in these release notes.